+++ skill-db/backend/src/main/java/com/cofinprobootcamp/backend/profile/CSVReader.java
@@ -16,7 +16,6 @@ import org.springframework.web.multipart.MultipartFile;
 import javax.validation.*;
 import java.io.IOException;
 import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.Set;
